Image history limit increased
The limit of messages from which images are sent to the api for every request has been increased from 3 to 8. This has come from a realization that the API is capable of handling it and 3 is too little to provide accurate context in certain situations, so 8 is a better balance.
This could be made a GUI settings option in a future revision.